requirements.txt
addict
cupy
future
lmdb
numpy>=1.17
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
torch>=1.9
torchvision
tqdm
wandb
yapf
timm>=0.6
ninja

